petpooja-voice-integration/requirements.txt
langchain
langchain-community
langchain-core
groq
python-dotenv
pydantic
faiss-cpu
sentence-transformers
tiktoken
chromadb
deepgram-sdk
openai-whisper
pyaudio
sounddevice
elevenlabs
twilio
langdetect
deep-translator
fastapi
uvicorn
requests
numpy
pandas
tqdm
orjson
motor
websockets
langchain-groq
python-multipart
certifi